UI changes & maths template & add Median/SD to Zoomer
Separate the main UI to 3 tab widget: user, setup and debug. Changes the maths function in maths.h to template to fit all format of data type. Delete the maths.cpp because the template can not be compile separately with the header file. Add the calcul and display of median&sd for the rectangle shape selected in Zoomer.